feat!: v7.36.0 — remove the Gemini Code Assist provider, repoint gemini-* at Antigravity
The v7.30.0 split shipped the `antigravity` provider and deprecated the `go@`
PREFIX, but never repointed the routing chain or retired the dead provider:

    "gemini-*": ["gemini-codeassist", "google", "openrouter"]

Google retired "Code Assist for individuals" for gemini-cli's OAuth client
(UNSUPPORTED_CLIENT), so the head of that chain could not authenticate for any
consumer account. Every bare `gemini-*` name paid a guaranteed-failing round-trip
and then fell through to the METERED `google` API — silently billing per-token
for a model the user's Antigravity subscription already covers. The chain is now
["antigravity", "google", "openrouter"], which is the subscription-first
convention this table's own header documents and every other family follows.

The provider is removed outright, not merely deregistered: definition, transport,
credential provider, OAuth registration, quota adapter, probe entry, telemetry
name, picker prefix, and the --gemini-login/--gemini-logout flags. Its
oauth-registry entries are gone too, so a leftover ~/.claudish/gemini-oauth.json
no longer reads as a live credential — that stale file is why a dead provider
kept its place in the config TUI's Test All list and failed every run.

auth/gemini-oauth.ts was a MIXED module: deleting it wholesale would have taken
ag@ down with it. The Antigravity half (project/tier resolution, retrieveUserQuota,
live served-set discovery) is extracted first to auth/antigravity-user.ts, then
the original is deleted. retrieveUserQuota deliberately keeps the gemini-cli
User-Agent it was written with: that exact call is verified working on Ultra and
loadCodeAssist is known to gate on identity, so changing it is a live experiment,
not a tidy-up.

`claudish quota gemini` now reports Antigravity rather than erroring on a product
that no longer exists; `go@` remains a deprecated alias → ag@.

Verified: typecheck + build clean; 706/708 provider+auth tests pass (the 1 failure
is the pre-existing 1Password non-interactive timeout, reproduced on main);
runtime check confirms the chain, the absent provider, and 29 providers registered.
The e2e suite shows the new chain live: "[Route] 2 providers for
gemini-3.1-pro-preview: Antigravity → OpenRouter".

Backend follow-up (probeModels still serves a gemini-codeassist row, plus a
newly-found family→served-id catalog gap) is written up for models-index in
TASK_retire_gemini_codeassist_probe.md.


Crafted with agentic harness Magus (https://github.com/MadAppGang/magus)
