+++ docs/api-key-architecture.md
@@ -31,7 +31,8 @@ When a model ID is provided, it's resolved in this order:
 | Prefix | Provider | API Key Env Var | Notes |
 |--------|----------|-----------------|-------|
 | `g/`, `gemini/` | Google Gemini | `GEMINI_API_KEY` | Direct Gemini API |
-| `go/` | Gemini Code Assist | OAuth | Requires `claudish --gemini-login` |
+| `ag/`, `antigravity/` | Antigravity (Gemini subscription) | OAuth | Requires `claudish login antigravity` |
+| `go/` | _deprecated alias → `ag/`_ | OAuth | Gemini Code Assist was retired by Google |
 | `v/`, `vertex/` | Vertex AI | `VERTEX_API_KEY` or `VERTEX_PROJECT` (OAuth) | Google Cloud |
 | `oai/` | OpenAI | `OPENAI_API_KEY` | Direct OpenAI API |
 | `mmax/`, `mm/` | MiniMax | `MINIMAX_API_KEY` | Anthropic-compatible |
