--- packages/cli/src/auth/gemini-oauth.ts
@@ -1,1346 +1,0 @@
-/**
- * Gemini OAuth Authentication Manager
- *
- * Handles OAuth2 PKCE flow for Gemini Code Assist API access.
- * Supports:
- * - Browser-based OAuth login with local callback server
- * - Secure credential storage with 0600 permissions
- * - Automatic token refresh with 5-minute buffer
- * - Singleton pattern for shared token management
- *
- * Credentials stored at: ~/.claudish/gemini-oauth.json
- */
-
-import { exec } from "node:child_process";
-import { createHash, randomBytes } from "node:crypto";
-import { closeSync, existsSync, openSync, readFileSync, unlinkSync, writeSync } from "node:fs";
-import { type IncomingMessage, type ServerResponse, createServer } from "node:http";
-import { homedir } from "node:os";
-import { join } from "node:path";
-import { promisify } from "node:util";
-import { log } from "../logger.js";
-
-const execAsync = promisify(exec);
-
-/**
- * OAuth credentials structure
- */
-export interface GeminiCredentials {
-  access_token: string;
-  refresh_token: string;
-  expires_at: number; // Unix timestamp (ms)
-}
-
-/**
- * Google OAuth token response
- */
-interface TokenResponse {
-  access_token: string;
-  refresh_token?: string;
-  expires_in: number;
-  token_type: string;
-}
-
-/**
- * Default OAuth credentials (Google's public OAuth client - same as gemini-cli)
- * These are PUBLIC credentials designed to be embedded in client applications.
- * Split to avoid false-positive secret scanning (GitHub detects base64 too).
- */
-const getDefaultClientId = (): string => {
-  // Public client ID from gemini-cli, split to avoid detection
-  const parts = [
-    "681255809395",
-    "oo8ft2oprdrnp9e3aqf6av3hmdib135j",
-    "apps",
-    "googleusercontent",
-    "com",
-  ];
-  return `${parts[0]}-${parts[1]}.${parts[2]}.${parts[3]}.${parts[4]}`;
-};
-const getDefaultClientSecret = (): string => {
-  // Public client secret from gemini-cli, split to avoid detection
-  const p = ["GOCSPX", "4uHgMPm", "1o7Sk", "geV6Cu5clXFsxl"];
-  return `${p[0]}-${p[1]}-${p[2]}-${p[3]}`;
-};
-
-/**
- * OAuth configuration (using Google's public OAuth client - same as gemini-cli)
- * Client ID/Secret can be overridden via environment variables if needed.
- */
-const OAUTH_CONFIG = {
-  clientId: process.env.GEMINI_CLIENT_ID || getDefaultClientId(),
-  clientSecret: process.env.GEMINI_CLIENT_SECRET || getDefaultClientSecret(),
-  authUrl: "https://accounts.google.com/o/oauth2/v2/auth",
-  tokenUrl: "https://oauth2.googleapis.com/token",
-  // redirectUri is built dynamically with the actual port
-  scopes: [
-    "https://www.googleapis.com/auth/cloud-platform",
-    "https://www.googleapis.com/auth/userinfo.email",
-    "https://www.googleapis.com/auth/userinfo.profile",
-  ],
-};
-
-/**
- * Client identity presented to the Code Assist backend.
- *
- * Presenting the Antigravity identity (`User-Agent` + `metadata.ideType`)
- * changes ONLY what `loadCodeAssist` reports — it makes the individuals/Ultra
- * tier visible. Verified 2026-08-02 by MITM-capturing the real Antigravity CLI
- * and replaying against claudish's own `claudish login gemini` token:
- *
- *   ideType=GEMINI_CLI  + GeminiCLI UA    → free-tier: UNSUPPORTED_CLIENT
- *   ideType=ANTIGRAVITY + antigravity UA  → currentTier free-tier, paidTier g1-ultra
- *   ideType=ANTIGRAVITY + GeminiCLI UA    → UNSUPPORTED_CLIENT  (UA matters too)
- *
- * BUT the actual generation call does NOT accept the spoof: streamGenerateContent
- * returns 403 PERMISSION_DENIED for every envelope and on both hosts. The
- * metadata endpoint gates on the identity headers; the generation endpoint gates
- * on the OAuth CLIENT that minted the token, which headers cannot fake. Reaching
- * generation needs an Antigravity-issued token — i.e. logging in with
- * Antigravity's OAuth client via GEMINI_CLIENT_ID / GEMINI_CLIENT_SECRET, a
- * user-supplied choice claudish does not ship.
- *
- * So this is DEFAULT OFF. On by default it would resolve every user to the Ultra
- * tier and then 403 on generation — strictly worse than the gemini-cli identity,
- * which at least reaches standard-tier with a project. Opt in with
- * CLAUDISH_GEMINI_ANTIGRAVITY=1 to inspect the tier (and to experiment once an
- * Antigravity-issued token is in play).
- */
-const USE_ANTIGRAVITY_GEMINI_IDENTITY = process.env.CLAUDISH_GEMINI_ANTIGRAVITY === "1";
-const USE_LEGACY_GEMINI_IDENTITY = !USE_ANTIGRAVITY_GEMINI_IDENTITY;
-
-/** The `metadata.ideType` sent on loadCodeAssist / onboardUser. */
-export const GEMINI_IDE_TYPE = USE_LEGACY_GEMINI_IDENTITY ? "GEMINI_CLI" : "ANTIGRAVITY";
-
-/** True when presenting as Antigravity (the default). */
-export const GEMINI_IDENTITY_IS_ANTIGRAVITY = !USE_LEGACY_GEMINI_IDENTITY;
-
-/**
- * Build the User-Agent for Code Assist requests. Antigravity's own UA carries no
- * model segment; the legacy gemini-cli UA does. Shared by the credential
- * provider (primary request headers) and the transport (429-fallback headers) so
- * both present one consistent identity.
- */
-export function buildCodeAssistUserAgent(model?: string): string {
-  if (USE_LEGACY_GEMINI_IDENTITY) {
-    const modelSegment = model || "gemini-code-assist";
-    return `GeminiCLI/0.5.6/${modelSegment} (${process.platform}; ${process.arch})`;
-  }
-  // Matches the captured Antigravity CLI exactly (os_type/arch tokens included).
-  return `antigravity/cli/1.1.9 (aidev_client; os_type=${process.platform}; arch=${process.arch}; auth_method=consumer)`;
-}
-
-/**
- * Manages OAuth authentication for Gemini Code Assist API
- */
-export class GeminiOAuth {
-  private static instance: GeminiOAuth | null = null;
-  private credentials: GeminiCredentials | null = null;
-  private refreshPromise: Promise<string> | null = null;
-  private tokenRefreshMargin = 5 * 60 * 1000; // Refresh 5 minutes before expiry
-  private oauthState: string | null = null; // CSRF protection
-
-  /**
-   * Get singleton instance
-   */
-  static getInstance(): GeminiOAuth {
-    if (!GeminiOAuth.instance) {
-      GeminiOAuth.instance = new GeminiOAuth();
-    }
-    return GeminiOAuth.instance;
-  }
-
-  /**
-   * Private constructor (singleton pattern)
-   */
-  private constructor() {
-    // Try to load existing credentials on startup
-    this.credentials = this.loadCredentials();
-  }
-
-  /**
-   * Re-read the credential file from disk into the in-memory singleton.
-   *
-   * The singleton loads credentials ONCE in its constructor, so a login that
-   * happens in a CHILD process (e.g. the config TUI spawns `claudish login
-   * gemini`) writes a fresh token file that this long-lived parent process
-   * never sees — its `getAccessToken()` keeps using the stale startup snapshot
-   * and fails until the whole process is relaunched. Calling this after a login
-   * child returns picks up the new token in-process, no restart needed.
-   */
-  reloadCredentials(): void {
-    this.credentials = this.loadCredentials();
-    // Drop any in-flight refresh bound to the old token.
-    this.refreshPromise = null;
-  }
-
-  /**
-   * Check if credentials exist (without validating expiry)
-   * Use this to determine if login is needed before making requests
-   */
-  hasCredentials(): boolean {
-    return this.credentials !== null && !!this.credentials.refresh_token;
-  }
-
-  /**
-   * Get credentials file path
-   */
-  private getCredentialsPath(): string {
-    const claudishDir = join(homedir(), ".claudish");
-    return join(claudishDir, "gemini-oauth.json");
-  }
-
-  /**
-   * Start OAuth login flow
-   * Opens browser, starts local callback server, exchanges code for tokens
-   */
-  async login(): Promise<void> {
-    log("[GeminiOAuth] Starting OAuth login flow");
-
-    // Generate PKCE verifier and challenge
-    const codeVerifier = this.generateCodeVerifier();
-    const codeChallenge = await this.generateCodeChallenge(codeVerifier);
-
-    // Generate state for CSRF protection
-    this.oauthState = randomBytes(32).toString("base64url");
-
-    // Start local callback server (uses random port) and wait for auth code
-    const { authCode, redirectUri } = await this.startCallbackServer(
-      codeChallenge,
-      this.oauthState
-    );
-
-    // Exchange auth code for tokens
-    const tokens = await this.exchangeCodeForTokens(authCode, codeVerifier, redirectUri);
-
-    // Save credentials
-    const credentials: GeminiCredentials = {
-      access_token: tokens.access_token,
-      refresh_token: tokens.refresh_token!,
-      expires_at: Date.now() + tokens.expires_in * 1000,
-    };
-
-    this.saveCredentials(credentials);
-    this.credentials = credentials;
-
-    // Clear state after successful login
-    this.oauthState = null;
-
-    log("[GeminiOAuth] Login successful");
-  }
-
-  /**
-   * Logout - delete stored credentials
-   */
-  async logout(): Promise<void> {
-    const credPath = this.getCredentialsPath();
-
-    if (existsSync(credPath)) {
-      unlinkSync(credPath);
-      log("[GeminiOAuth] Credentials deleted");
-    }
-
-    this.credentials = null;
-  }
-
-  /**
-   * Get valid access token, refreshing if needed
-   */
-  async getAccessToken(): Promise<string> {
-    // If refresh already in progress, wait for it
-    if (this.refreshPromise) {
-      log("[GeminiOAuth] Waiting for in-progress refresh");
-      return this.refreshPromise;
-    }
-
-    // Check if we have credentials
-    if (!this.credentials) {
-      throw new Error(
-        "No Gemini OAuth credentials found. Please run `claudish login gemini` first."
-      );
-    }
-
-    // Check if token is still valid
-    if (this.isTokenValid()) {
-      return this.credentials.access_token;
-    }
-
-    // Start refresh (lock to prevent duplicate refreshes)
-    this.refreshPromise = this.doRefreshToken();
-
-    try {
-      const token = await this.refreshPromise;
-      return token;
-    } finally {
-      this.refreshPromise = null;
-    }
-  }
-
-  /**
-   * Force refresh the access token
-   */
-  async refreshToken(): Promise<void> {
-    if (!this.credentials) {
-      throw new Error(
-        "No Gemini OAuth credentials found. Please run `claudish login gemini` first."
-      );
-    }
-
-    await this.doRefreshToken();
-  }
-
-  /**
-   * Check if cached token is still valid
-   */
-  private isTokenValid(): boolean {
-    if (!this.credentials) return false;
-    return Date.now() < this.credentials.expires_at - this.tokenRefreshMargin;
-  }
-
-  /**
-   * Perform the actual token refresh
-   */
-  private async doRefreshToken(): Promise<string> {
-    if (!this.credentials) {
-      throw new Error(
-        "No Gemini OAuth credentials found. Please run `claudish login gemini` first."
-      );
-    }
-
-    log("[GeminiOAuth] Refreshing access token");
-
-    try {
-      const response = await fetch(OAUTH_CONFIG.tokenUrl, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-        body: new URLSearchParams({
-          grant_type: "refresh_token",
-          refresh_token: this.credentials.refresh_token,
-          client_id: OAUTH_CONFIG.clientId,
-          client_secret: OAUTH_CONFIG.clientSecret,
-        }),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`Token refresh failed: ${response.status} - ${errorText}`);
-      }
-
-      const tokens = (await response.json()) as TokenResponse;
-
-      // Update credentials (keep existing refresh token if new one not provided)
-      const updatedCredentials: GeminiCredentials = {
-        access_token: tokens.access_token,
-        refresh_token: tokens.refresh_token || this.credentials.refresh_token,
-        expires_at: Date.now() + tokens.expires_in * 1000,
-      };
-
-      this.saveCredentials(updatedCredentials);
-      this.credentials = updatedCredentials;
-
-      log(
-        `[GeminiOAuth] Token refreshed, valid until ${new Date(updatedCredentials.expires_at).toISOString()}`
-      );
-
-      return updatedCredentials.access_token;
-    } catch (e: any) {
-      log(`[GeminiOAuth] Refresh failed: ${e.message}`);
-      throw new Error(
-        `OAuth credentials invalid. Please run \`claudish login gemini\` again.\n\nDetails: ${e.message}`
-      );
-    }
-  }
-
-  /**
-   * Load credentials from file
-   */
-  private loadCredentials(): GeminiCredentials | null {
-    const credPath = this.getCredentialsPath();
-
-    if (!existsSync(credPath)) {
-      return null;
-    }
-
-    try {
-      const data = readFileSync(credPath, "utf-8");
-      const credentials = JSON.parse(data) as GeminiCredentials;
-
-      // Validate structure
-      if (!credentials.access_token || !credentials.refresh_token || !credentials.expires_at) {
-        log("[GeminiOAuth] Invalid credentials file structure");
-        return null;
-      }
-
-      log("[GeminiOAuth] Loaded credentials from file");
-      return credentials;
-    } catch (e: any) {
-      log(`[GeminiOAuth] Failed to load credentials: ${e.message}`);
-      return null;
-    }
-  }
-
-  /**
-   * Save credentials to file with 0600 permissions
-   */
-  private saveCredentials(credentials: GeminiCredentials): void {
-    const credPath = this.getCredentialsPath();
-    const claudishDir = join(homedir(), ".claudish");
-
-    // Ensure directory exists
-    if (!existsSync(claudishDir)) {
-      const { mkdirSync } = require("node:fs");
-      mkdirSync(claudishDir, { recursive: true });
-    }
-
-    // Atomically create file with secure permissions (0600) to prevent race condition
-    const fd = openSync(credPath, "w", 0o600);
-    try {
-      const data = JSON.stringify(credentials, null, 2);
-      writeSync(fd, data, 0, "utf-8");
-    } finally {
-      closeSync(fd);
-    }
-
-    log(`[GeminiOAuth] Credentials saved to ${credPath}`);
-  }
-
-  /**
-   * Generate PKCE code verifier (random 128-character string)
-   */
-  private generateCodeVerifier(): string {
-    return randomBytes(64).toString("base64url");
-  }
-
-  /**
-   * Generate PKCE code challenge (SHA256 hash of verifier)
-   */
-  private async generateCodeChallenge(verifier: string): Promise<string> {
-    const hash = createHash("sha256").update(verifier).digest("base64url");
-    return hash;
-  }
-
-  /**
-   * Build OAuth authorization URL
-   */
-  private buildAuthUrl(codeChallenge: string, state: string, redirectUri: string): string {
-    const params = new URLSearchParams({
-      client_id: OAUTH_CONFIG.clientId,
-      redirect_uri: redirectUri,
-      response_type: "code",
-      scope: OAUTH_CONFIG.scopes.join(" "),
-      code_challenge: codeChallenge,
-      code_challenge_method: "S256",
-      access_type: "offline", // Request refresh token
-      prompt: "consent", // Force consent screen to get refresh token
-      state, // CSRF protection
-    });
-
-    return `${OAUTH_CONFIG.authUrl}?${params.toString()}`;
-  }
-
-  /**
-   * Start local callback server and wait for authorization code
-   * Uses random available port (port 0) to avoid conflicts
-   */
-  private async startCallbackServer(
-    codeChallenge: string,
-    state: string
-  ): Promise<{ authCode: string; redirectUri: string }> {
-    return new Promise((resolve, reject) => {
-      let redirectUri = "";
-
-      const server = createServer((req: IncomingMessage, res: ServerResponse) => {
-        const url = new URL(req.url!, redirectUri.replace("/callback", ""));
-
-        if (url.pathname === "/callback") {
-          const code = url.searchParams.get("code");
-          const callbackState = url.searchParams.get("state");
-          const error = url.searchParams.get("error");
-
-          if (error) {
-            res.writeHead(400, { "Content-Type": "text/html" });
-            res.end(`
-              <html>
-                <body>
-                  <h1>Authentication Failed</h1>
-                  <p>Error: ${error}</p>
-                  <p>You can close this window.</p>
-                </body>
-              </html>
-            `);
-            server.close();
-            reject(new Error(`OAuth error: ${error}`));
-            return;
-          }
-
-          // Validate state parameter (CSRF protection)
-          if (!callbackState || callbackState !== this.oauthState) {
-            res.writeHead(400, { "Content-Type": "text/html" });
-            res.end(`
-              <html>
-                <body>
-                  <h1>Authentication Failed</h1>
-                  <p>Invalid state parameter. Possible CSRF attack.</p>
-                  <p>You can close this window.</p>
-                </body>
-              </html>
-            `);
-            server.close();
-            reject(new Error("Invalid OAuth state parameter (CSRF protection)"));
-            return;
-          }
-
-          if (!code) {
-            res.writeHead(400, { "Content-Type": "text/html" });
-            res.end(`
-              <html>
-                <body>
-                  <h1>Authentication Failed</h1>
-                  <p>No authorization code received.</p>
-                  <p>You can close this window.</p>
-                </body>
-              </html>
-            `);
-            server.close();
-            reject(new Error("No authorization code received"));
-            return;
-          }
-
-          // Success
-          res.writeHead(200, { "Content-Type": "text/html" });
-          res.end(`
-            <html>
-              <body>
-                <h1>Authentication Successful!</h1>
-                <p>You can now close this window and return to your terminal.</p>
-              </body>
-            </html>
-          `);
-
-          server.close();
-          resolve({ authCode: code, redirectUri });
-        } else {
-          res.writeHead(404, { "Content-Type": "text/plain" });
-          res.end("Not found");
-        }
-      });
-
-      // Listen on port 0 to get a random available port
-      server.listen(0, () => {
-        const address = server.address();
-        if (!address || typeof address === "string") {
-          reject(new Error("Failed to get server port"));
-          return;
-        }
-
-        const port = address.port;
-        redirectUri = `http://localhost:${port}/callback`;
-        log(`[GeminiOAuth] Callback server started on http://localhost:${port}`);
-
-        // Build auth URL with the actual port and open browser
-        const authUrl = this.buildAuthUrl(codeChallenge, state, redirectUri);
-        this.openBrowser(authUrl);
-      });
-
-      server.on("error", (err) => {
-        reject(new Error(`Failed to start callback server: ${err.message}`));
-      });
-
-      // Timeout after 5 minutes
-      setTimeout(
-        () => {
-          server.close();
-          reject(new Error("OAuth login timed out after 5 minutes"));
-        },
-        5 * 60 * 1000
-      );
-    });
-  }
-
-  /**
-   * Exchange authorization code for access/refresh tokens
-   */
-  private async exchangeCodeForTokens(
-    code: string,
-    verifier: string,
-    redirectUri: string
-  ): Promise<TokenResponse> {
-    log("[GeminiOAuth] Exchanging auth code for tokens");
-
-    try {
-      const response = await fetch(OAUTH_CONFIG.tokenUrl, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-        body: new URLSearchParams({
-          grant_type: "authorization_code",
-          code,
-          redirect_uri: redirectUri,
-          client_id: OAUTH_CONFIG.clientId,
-          client_secret: OAUTH_CONFIG.clientSecret,
-          code_verifier: verifier,
-        }),
-      });
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(`Token exchange failed: ${response.status} - ${errorText}`);
-      }
-
-      const tokens = (await response.json()) as TokenResponse;
-
-      if (!tokens.access_token || !tokens.refresh_token) {
-        throw new Error("Token response missing access_token or refresh_token");
-      }
-
-      return tokens;
-    } catch (e: any) {
-      throw new Error(`Failed to authenticate with Google OAuth: ${e.message}`);
-    }
-  }
-
-  /**
-   * Open URL in default browser
-   */
-  private async openBrowser(url: string): Promise<void> {
-    const platform = process.platform;
-
-    try {
-      if (platform === "darwin") {
-        await execAsync(`open "${url}"`);
-      } else if (platform === "win32") {
-        await execAsync(`start "${url}"`);
-      } else {
-        // Linux/Unix
-        await execAsync(`xdg-open "${url}"`);
-      }
-
-      console.log("\nOpening browser for authentication...");
-      console.log(`If the browser doesn't open, visit this URL:\n${url}\n`);
-    } catch (e: any) {
-      console.log("\nPlease open this URL in your browser to authenticate:");
-      console.log(url);
-      console.log("");
-    }
-  }
-}
-
-/**
- * Get the shared GeminiOAuth instance
- */
-export function getGeminiOAuth(): GeminiOAuth {
-  return GeminiOAuth.getInstance();
-}
-
-/**
- * Pick up a Gemini login that happened in a child process: re-read the token
- * file into the singleton AND drop the cached Code Assist project/tier (a new
- * login may be a different account). Call after a `claudish login gemini` child
- * returns so the running parent uses the fresh credential without a restart.
- *
- * NOTE: resetGeminiUserCache is defined later in this module but hoisted as a
- * function declaration, so referencing it here is safe.
- */
-export function reloadGeminiCredentials(): void {
-  GeminiOAuth.getInstance().reloadCredentials();
-  resetGeminiUserCache();
-}
-
-// ============================================================================
-// Code Assist User Setup Flow
-// ============================================================================
-
-const CODE_ASSIST_API_BASE = "https://cloudcode-pa.googleapis.com/v1internal";
-
-/**
- * Cold-start seed for the Code Assist capacity-fallback chain.
- *
- * Used ONLY before the first live `getServedCodeAssistModels()` succeeds (or
- * if the quota fetch fails). The PRIMARY source of "which models does this tier
- * serve" is the live served set read from `retrieveUserQuota` buckets, because
- * the subset depends on the account's tier AND changes over time: 3.x
- * pro/flash-preview once worked on this backend and now 404, and models like
- * gemini-3.6-flash are direct-API-only. A hardcoded roster rots; this seed just
- * keeps the 404 hint and the fallback chain sensible during cold start /
- * quota-fetch failure.
- *
- * Kept current with the known-good served set as of the last manual check;
- * drift self-corrects once the first request fetches quota. Sorted at use by
- * `rankCodeAssistModel`, so order here is incidental.
- */
-export const CODE_ASSIST_FALLBACK_CHAIN = [
-  "gemini-2.5-pro",
-  "gemini-2.5-flash",
-  "gemini-2.5-flash-lite",
-  "gemini-3.1-flash-lite",
-] as const;
-
-export type CodeAssistFallbackModel = (typeof CODE_ASSIST_FALLBACK_CHAIN)[number];
-
-/**
- * Preference rank for a Code Assist model — a RULE, not a pinned roster.
- *
- * Lower number = tried first in capacity fallback / shown first in the quota
- * list. Ordering: pro → flash → flash-lite (then everything else), newest
- * version first within each tier. Operates on whatever the live served set
- * returns, so it survives tier changes without code edits.
- */
-export function rankCodeAssistModel(model: string): number {
-  const lower = model.toLowerCase();
-  let tier: number;
-  if (lower.includes("pro")) tier = 0;
-  else if (lower.includes("lite")) tier = 2;
-  else if (lower.includes("flash")) tier = 1;
-  else tier = 3;
-  // Anchor the version to the family prefix. A bare leftmost-number scan reads
-  // the date suffix in IDs like "gemini-exp-1206" as version 1206, which sorts
-  // that model into the middle of a tier it does not belong to.
-  const vMatch = lower.match(/^gemini-(\d+(?:\.\d+)?)(?![\d.])/);
-  const version = vMatch ? Number.parseFloat(vMatch[1]) : 0;
-  // 1000 spacing keeps tiers from bleeding together across plausible versions.
-  return tier * 1000 - version;
-}
-
-let servedModelsCache: string[] | null = null;
-let servedModelsCacheAt = 0;
-const SERVED_MODELS_TTL_MS = 10 * 60 * 1000;
-
-/**
- * Discover which models this account's Code Assist tier currently serves — LIVE.
- *
- * `retrieveUserQuota` returns one bucket per served model, so its bucket list
- * IS the authoritative served set. We read it instead of hardcoding a roster.
- * Cached for SERVED_MODELS_TTL_MS: the served set changes on the daily-quota
- * cadence, not per request, and `retrieveUserQuota` is a real network call we
- * don't want on every refreshAuth.
- *
- * Falls back to CODE_ASSIST_FALLBACK_CHAIN (the cold-start seed) when the quota
- * fetch is unavailable, so the 404 hint and capacity fallback degrade gracefully.
- *
- * NOTE: the backend also exposes `v1internal:fetchAvailableModels`, which is a
- * purpose-built served-set endpoint rather than an inference from quota buckets.
- * Worth migrating to once its request shape is pinned down — see ROADMAP.
- */
-export async function getServedCodeAssistModels(
-  accessToken: string,
-  projectId: string,
-  opts?: { force?: boolean }
-): Promise<string[]> {
-  const now = Date.now();
-  if (!opts?.force && servedModelsCache && now - servedModelsCacheAt < SERVED_MODELS_TTL_MS) {
-    return servedModelsCache;
-  }
-  try {
-    const data = await retrieveUserQuota(accessToken, projectId);
-    const ids = (data?.buckets ?? [])
-      .map((b) => b.modelId)
-      .filter((m): m is string => typeof m === "string" && m.length > 0);
-    if (ids.length > 0) {
-      const sorted = ids.slice().sort((a, b) => rankCodeAssistModel(a) - rankCodeAssistModel(b));
-      servedModelsCache = sorted;
-      servedModelsCacheAt = now;
-      return sorted;
-    }
-  } catch (err) {
-    log(`[GeminiOAuth] getServedCodeAssistModels error: ${err}`);
-  }
-  if (servedModelsCache) return servedModelsCache;
-  return CODE_ASSIST_FALLBACK_CHAIN.slice().sort(
-    (a, b) => rankCodeAssistModel(a) - rankCodeAssistModel(b)
-  );
-}
-
-/** Test seam: clear the served-models cache between tests. */
-export function _resetServedModelsCache(): void {
-  servedModelsCache = null;
-  servedModelsCacheAt = 0;
-}
-
-interface ClientMetadata {
-  // Optional because the Antigravity identity sends ONLY ideType — the free tier
-  // returns the project inline without pluginType/platform, and adding them on
-  // that path is unproven.
-  pluginType?: string;
-  ideType: string;
-  platform?: string;
-  duetProject?: string;
-}
-
-interface AllowedTier {
-  id: string;
-  displayName?: string;
-  name?: string;
-  /**
-   * True when the caller must supply the GCP project themselves. Google
-   * auto-provisions a project for the individuals tier but not for seat-based
-   * tiers, so onboarding without a project silently yields no project ID.
-   */
-  userDefinedCloudaicompanionProject?: boolean;
-}
-
-/**
- * A tier the backend refuses to onboard this caller to, and why.
- *
- * This is where Google explains an onboarding refusal — e.g. the individuals
- * tier answering UNSUPPORTED_CLIENT with a migration instruction. Dropping it
- * is what turned a specific, actionable verdict into the opaque "no project ID
- * returned" error, so it is captured and surfaced.
- */
-interface IneligibleTier {
-  tierId?: string;
-  tierName?: string;
-  reasonCode?: string;
-  reasonMessage?: string;
-}
-
-interface LoadCodeAssistResponse {
-  currentTier?: string | { id?: string };
-  paidTier?: { id?: string; name?: string };
-  cloudaicompanionProject?: string;
-  allowedTiers?: AllowedTier[];
-  ineligibleTiers?: IneligibleTier[];
-}
-
-interface LROResponse {
-  done?: boolean;
-  error?: { code: number; message: string };
-  response?: {
-    cloudaicompanionProject?: { id: string };
-  };
-}
-
-/**
- * Get a valid access token (refreshing if needed)
- * Helper function for handlers to use
- */
-export async function getValidAccessToken(): Promise<string> {
-  const oauth = GeminiOAuth.getInstance();
-  return oauth.getAccessToken();
-}
-
-// Cache for project ID and tier to avoid setup on every request
-let cachedProjectId: string | null = null;
-let cachedTierId: string | null = null;
-let cachedTierName: string | null = null;
-/**
- * A terminal setup failure (message only), latched so retries fail instantly.
- * Only configuration verdicts land here — never transient network errors, which
- * must stay retryable.
- */
-let cachedSetupError: string | null = null;
-
-/**
- * Clear the cached Code Assist project/tier (set by setupGeminiUser). Must run
- * on re-login: a fresh login can be a DIFFERENT Google account with a different
- * project/tier, and the stale cache would otherwise be sent in the request
- * envelope (wrong project → 4xx). Called by reloadGeminiCredentials().
- */
-export function resetGeminiUserCache(): void {
-  cachedProjectId = null;
-  cachedTierId = null;
-  cachedTierName = null;
-  cachedSetupError = null;
-}
-
-/** Short display names for known tier IDs (status bar needs compact names) */
-const TIER_SHORT_NAMES: Record<string, string> = {
-  "free-tier": "GeminiCA Free",
-  "standard-tier": "GeminiCA Std",
-  "g1-pro-tier": "GeminiCA Pro",
-  "legacy-tier": "GeminiCA Legacy",
-};
-
-/**
- * Get a compact display name for the status bar.
- * Returns short names like "GeminiCA Std", "GeminiCA Pro".
- *
- * Before the tier is resolved we do NOT guess "Free": that label was printed on
- * every pre-auth error line, telling users they were on a tier this account may
- * not even be eligible for. An unresolved tier is unknown, not free.
- */
-export function getGeminiTierDisplayName(): string {
-  if (!cachedTierId) return "GeminiCA";
-  return TIER_SHORT_NAMES[cachedTierId] || cachedTierId.replace(/-tier$/, "");
-}
-
-/**
- * Get the full tier name from the API (for quota command / detailed views).
- */
-export function getGeminiTierFullName(): string {
-  if (cachedTierName) return cachedTierName;
-  return getGeminiTierDisplayName();
-}
-
-/**
- * Marker for auth failures that RETRYING CANNOT FIX.
- *
- * The generic auth-failure path answers 401 on purpose, so FallbackHandler can
- * advance the provider chain. That is right for a credential that might come
- * back (expired token, transient backend). It is wrong for a configuration
- * verdict: the client retried a missing-project error ~11 times over two
- * minutes of backoff, burying the actionable message under
- * "API error · Retrying". Callers check `terminal` and answer 400 instead, so
- * the explanation reaches the user immediately.
- */
-export function isTerminalGeminiSetupError(err: unknown): boolean {
-  return Boolean(err && typeof err === "object" && (err as { terminal?: unknown }).terminal);
-}
-
-function makeTerminalSetupError(message: string): Error {
-  const err = new Error(message);
-  (err as Error & { terminal?: boolean }).terminal = true;
-  return err;
-}
-
-/**
- * Explain, in the user's terms, why Code Assist onboarding cannot produce a
- * project — and what to do about it.
- *
- * Two facts drive every instance of this failure and both come straight from
- * loadCodeAssist: the tier we were steered to needs a caller-supplied project,
- * and (usually) the tier that would NOT have needed one was refused with a
- * reason. Quoting Google's own reasonMessage matters because it is the only
- * place a migration instruction or entitlement change is ever stated.
- */
-function buildProjectRequiredError(
-  tierId: string,
-  chosenTier: AllowedTier | undefined,
-  loadRes: LoadCodeAssistResponse
-): string {
-  const tierLabel = chosenTier?.name || chosenTier?.displayName || tierId;
-  const lines = [
-    `Gemini Code Assist requires a Google Cloud project for the "${tierLabel}" tier (${tierId}), and none is configured.`,
-    "",
-    "Set the project that holds your Code Assist license:",
-    "  export GOOGLE_CLOUD_PROJECT='your-project-id'",
-  ];
-
-  // Surface every refused tier verbatim. When the individuals/subscription tier
-  // is refused, this is the ONLY signal saying so — without it the user sees a
-  // project error and reasonably assumes claudish lost their subscription.
-  const refused = loadRes.ineligibleTiers ?? [];
-  if (refused.length > 0) {
-    lines.push("", "Google refused the other tier(s) on this account:");
-    for (const t of refused) {
-      const name = t.tierName || t.tierId || "unknown tier";
-      const code = t.reasonCode ? ` [${t.reasonCode}]` : "";
-      lines.push(`  • ${name}${code}`);
-      if (t.reasonMessage) lines.push(`    ${t.reasonMessage}`);
-    }
-  }
-
-  lines.push(
-    "",
-    "No project / no Code Assist seat? Use the direct Gemini API instead:",
-    "  export GEMINI_API_KEY='your-key'   # https://aistudio.google.com/app/apikey",
-    '  claudish --model google@gemini-2.5-pro "..."'
-  );
-
-  return lines.join("\n");
-}
-
-/**
- * Setup the Gemini user (loadCodeAssist + onboardUser flow)
- * Returns the projectId and tierId to use for requests.
- * Caches the result to avoid repeated API calls.
- */
-export async function setupGeminiUser(
-  accessToken: string
-): Promise<{ projectId: string; tierId: string }> {
-  // Return cached results if available
-  if (cachedProjectId && cachedTierId) {
-    log(`[GeminiOAuth] Using cached project ID: ${cachedProjectId}, tier: ${cachedTierId}`);
-    return { projectId: cachedProjectId, tierId: cachedTierId };
-  }
-
-  // A configuration failure (no project for a seat-based tier, tier refused) is
-  // deterministic for the life of the process — env vars do not change mid-run.
-  // The caller's 401 is retried by FallbackHandler, so without this latch one
-  // misconfiguration replayed loadCodeAssist + onboardUser on EVERY retry: ~11
-  // round trips of network for an outcome that was settled on the first.
-  if (cachedSetupError) {
-    log("[GeminiOAuth] Re-raising cached setup failure (no network retry)");
-    throw makeTerminalSetupError(cachedSetupError);
-  }
-
-  const envProject = process.env.GOOGLE_CLOUD_PROJECT || process.env.GOOGLE_CLOUD_PROJECT_ID;
-
-  // 1. loadCodeAssist - check if user is already set up
-  log("[GeminiOAuth] Calling loadCodeAssist...");
-  const loadRes = await callLoadCodeAssist(accessToken, envProject);
-  log(`[GeminiOAuth] loadCodeAssist response: ${JSON.stringify(loadRes)}`);
-
-  // Resolve tier: paidTier.id takes precedence over currentTier (matches gemini-cli)
-  const resolvedTier =
-    loadRes.paidTier?.id ||
-    (typeof loadRes.currentTier === "object" ? loadRes.currentTier?.id : loadRes.currentTier) ||
-    null;
-
-  if ((loadRes.currentTier || loadRes.paidTier) && loadRes.cloudaicompanionProject) {
-    const projectId = envProject || loadRes.cloudaicompanionProject;
-    if (projectId) {
-      cachedProjectId = projectId;
-      cachedTierId = resolvedTier || "free-tier";
-      cachedTierName = loadRes.paidTier?.name || null;
-      log(`[GeminiOAuth] User already set up, project: ${projectId}, tier: ${cachedTierId}`);
-      return { projectId, tierId: cachedTierId };
-    }
-  }
-
-  // 2. onboardUser - use the best tier available for this user
-  //    The server returns allowedTiers sorted by priority (best first).
-  //    Free tier must NOT send a project ID (Google provisions one).
-  //    Paid tiers (standard, legacy) require a project ID.
-  const tierId = resolvedTier || loadRes.allowedTiers?.[0]?.id || "free-tier";
-  const isFree = tierId === "free-tier";
-  const onboardProject = isFree ? undefined : envProject;
-
-  // Seat-based tiers require a caller-supplied project. Onboarding without one
-  // "succeeds" and returns no project ID, which used to surface as the opaque
-  // "onboarding completed but no project ID returned" AFTER up to 60s of
-  // polling. Refuse up front with the reason and the exact remedy.
-  const chosenTier = loadRes.allowedTiers?.find((t) => t.id === tierId);
-  if (!isFree && chosenTier?.userDefinedCloudaicompanionProject && !onboardProject) {
-    cachedSetupError = buildProjectRequiredError(tierId, chosenTier, loadRes);
-    throw makeTerminalSetupError(cachedSetupError);
-  }
-
-  const MAX_POLL_ATTEMPTS = 30; // 60 seconds max (30 * 2s)
-
-  log(`[GeminiOAuth] Onboarding user to ${tierId}...`);
-  let lro = await callOnboardUser(accessToken, tierId, onboardProject);
-  log(`[GeminiOAuth] Initial onboardUser response: done=${lro.done}`);
-
-  // Poll LRO until done (with timeout)
-  let attempts = 0;
-  while (!lro.done && attempts < MAX_POLL_ATTEMPTS) {
-    attempts++;
-    log(`[GeminiOAuth] Polling onboardUser (attempt ${attempts}/${MAX_POLL_ATTEMPTS})...`);
-    await new Promise((r) => setTimeout(r, 2000));
-    lro = await callOnboardUser(accessToken, tierId, onboardProject);
-  }
-
-  if (!lro.done) {
-    throw new Error(`Gemini onboarding timed out after ${MAX_POLL_ATTEMPTS * 2} seconds`);
-  }
-
-  if (lro.error) {
-    throw new Error(`Gemini onboarding failed: ${JSON.stringify(lro.error)}`);
-  }
-
-  const projectId = lro.response?.cloudaicompanionProject?.id;
-  if (!projectId) {
-    if (envProject) {
-      cachedProjectId = envProject;
-      cachedTierId = tierId;
-      return { projectId: envProject, tierId };
-    }
-    cachedSetupError = buildProjectRequiredError(tierId, chosenTier, loadRes);
-    throw makeTerminalSetupError(cachedSetupError);
-  }
-
-  cachedProjectId = projectId;
-  cachedTierId = tierId;
-  log(`[GeminiOAuth] Onboarding complete, project: ${projectId}, tier: ${tierId}`);
-  return { projectId, tierId };
-}
-
-async function callLoadCodeAssist(
-  accessToken: string,
-  projectId?: string
-): Promise<LoadCodeAssistResponse> {
-  // Match the captured request shape per identity. The Antigravity flow sends
-  // ONLY {ideType} — adding pluginType/platform is unproven on that path and the
-  // free tier returns the project inline without them, so keep it minimal.
-  const metadata: ClientMetadata = GEMINI_IDENTITY_IS_ANTIGRAVITY
-    ? { ideType: GEMINI_IDE_TYPE }
-    : {
-        pluginType: "GEMINI",
-        ideType: GEMINI_IDE_TYPE,
-        platform: "PLATFORM_UNSPECIFIED",
-        duetProject: projectId,
-      };
-
-  const res = await fetch(`${CODE_ASSIST_API_BASE}:loadCodeAssist`, {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      "Content-Type": "application/json",
-      "User-Agent": buildCodeAssistUserAgent(),
-    },
-    body: JSON.stringify({ metadata, cloudaicompanionProject: projectId }),
-  });
-
-  if (!res.ok) {
-    throw new Error(`loadCodeAssist failed: ${res.status} ${await res.text()}`);
-  }
-
-  return (await res.json()) as LoadCodeAssistResponse;
-}
-
-async function callOnboardUser(
-  accessToken: string,
-  tierId: string,
-  projectId?: string
-): Promise<LROResponse> {
-  const metadata: ClientMetadata = {
-    pluginType: "GEMINI",
-    ideType: GEMINI_IDE_TYPE,
-    platform: "PLATFORM_UNSPECIFIED",
-    duetProject: projectId,
-  };
-
-  const res = await fetch(`${CODE_ASSIST_API_BASE}:onboardUser`, {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      "Content-Type": "application/json",
-      "User-Agent": buildCodeAssistUserAgent(),
-    },
-    body: JSON.stringify({
-      tierId,
-      metadata,
-      cloudaicompanionProject: projectId,
-    }),
-  });
-
-  if (!res.ok) {
-    throw new Error(`onboardUser failed: ${res.status} ${await res.text()}`);
-  }
-
-  return (await res.json()) as LROResponse;
-}
-
-/** Quota bucket from retrieveUserQuota API */
-export interface QuotaBucket {
-  modelId?: string;
-  remainingFraction?: number;
-  remainingAmount?: string;
-  resetTime?: string;
-  tokenType?: string;
-}
-
-/**
- * Retrieve per-model quota usage from Code Assist API.
- * Returns quota buckets with remaining capacity per model.
- * Uses cached projectId and accessToken — call after setupGeminiUser.
- */
-export async function retrieveUserQuota(
-  accessToken: string,
-  projectId: string
-): Promise<{ buckets?: QuotaBucket[] } | null> {
-  try {
-    const res = await fetch(`${CODE_ASSIST_API_BASE}:retrieveUserQuota`, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-        "Content-Type": "application/json",
-        "User-Agent": `GeminiCLI/0.5.6/gemini-code-assist (${process.platform}; ${process.arch})`,
-      },
-      body: JSON.stringify({ project: projectId }),
-    });
-    if (!res.ok) {
-      log(`[GeminiOAuth] retrieveUserQuota failed: ${res.status}`);
-      return null;
-    }
-    return (await res.json()) as { buckets?: QuotaBucket[] };
-  } catch (err) {
-    log(`[GeminiOAuth] retrieveUserQuota error: ${err}`);
-    return null;
-  }
-}
-
-// ============================================================================
-// Antigravity identity (ALWAYS on for the antigravity provider)
-// ============================================================================
-//
-// The `google` / `gemini-codeassist` (go@) path above gates the Antigravity
-// identity behind CLAUDISH_GEMINI_ANTIGRAVITY. The dedicated antigravity
-// provider (ag@) is DIFFERENT: it authenticates with an Antigravity-issued
-// OAuth token (read from the shared agy store), so it ALWAYS presents the
-// Antigravity identity — the UA + `metadata.ideType: ANTIGRAVITY` — regardless
-// of that env flag. With this identity `loadCodeAssist` reports the
-// individuals/Ultra tier and auto-provisions a project inline; no onboardUser
-// and no GOOGLE_CLOUD_PROJECT are needed.
-
-/** The `metadata.ideType` sent by the Antigravity provider. */
-const ANTIGRAVITY_IDE_TYPE = "ANTIGRAVITY";
-
-/**
- * Antigravity's own User-Agent — matches the captured Antigravity CLI exactly.
- * Unconditional (independent of CLAUDISH_GEMINI_ANTIGRAVITY): this provider IS
- * Antigravity.
- */
-export function buildAntigravityUserAgent(): string {
-  return `antigravity/cli/1.1.9 (aidev_client; os_type=${process.platform}; arch=${process.arch}; auth_method=consumer)`;
-}
-
-// Antigravity setup cache — SEPARATE from the gemini-cli one (setupGeminiUser),
-// because a different identity resolves a different project/tier.
-let cachedAgProjectId: string | null = null;
-let cachedAgTierId: string | null = null;
-let cachedAgTierName: string | null = null;
-
-/** Clear the cached Antigravity project/tier (e.g. on account switch). */
-export function resetAntigravityUserCache(): void {
-  cachedAgProjectId = null;
-  cachedAgTierId = null;
-  cachedAgTierName = null;
-  agServedCache = null;
-  agServedCacheAt = 0;
-}
-
-/** loadCodeAssist with the Antigravity identity (minimal `{ ideType }` metadata). */
-async function callLoadCodeAssistAntigravity(accessToken: string): Promise<LoadCodeAssistResponse> {
-  const res = await fetch(`${CODE_ASSIST_API_BASE}:loadCodeAssist`, {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      "Content-Type": "application/json",
-      "User-Agent": buildAntigravityUserAgent(),
-    },
-    // The Antigravity path sends ONLY ideType — the individuals tier returns the
-    // project inline without pluginType/platform.
-    body: JSON.stringify({ metadata: { ideType: ANTIGRAVITY_IDE_TYPE } }),
-  });
-
-  if (!res.ok) {
-    throw new Error(`loadCodeAssist (antigravity) failed: ${res.status} ${await res.text()}`);
-  }
-
-  return (await res.json()) as LoadCodeAssistResponse;
-}
-
-/**
- * Resolve the project + tier for the Antigravity flow.
- *
- * loadCodeAssist (Antigravity identity) auto-provisions `cloudaicompanionProject`
- * inline for the individuals/Ultra tier, so — unlike setupGeminiUser — there is
- * NO onboardUser poll and NO GOOGLE_CLOUD_PROJECT requirement. `paidTier.id`
- * wins over `currentTier` (matches the go@ logic). Cached for the life of the
- * process (env/account do not change mid-run).
- */
-export async function setupAntigravityUser(
-  accessToken: string
-): Promise<{ projectId: string; tierId: string }> {
-  if (cachedAgProjectId && cachedAgTierId) {
-    return { projectId: cachedAgProjectId, tierId: cachedAgTierId };
-  }
-
-  const loadRes = await callLoadCodeAssistAntigravity(accessToken);
-  log(`[Antigravity] loadCodeAssist response: ${JSON.stringify(loadRes)}`);
-
-  const resolvedTier =
-    loadRes.paidTier?.id ||
-    (typeof loadRes.currentTier === "object" ? loadRes.currentTier?.id : loadRes.currentTier) ||
-    "free-tier";
-
-  const projectId = loadRes.cloudaicompanionProject;
-  if (!projectId) {
-    // A terminal, non-retryable configuration verdict — surface it inline
-    // instead of looping through the client's retry/backoff.
-    throw makeTerminalSetupError(
-      "Antigravity did not return a project for this account. Sign in to Antigravity " +
-        "(the `agy` CLI) and try again, or use g@<model> with GEMINI_API_KEY " +
-        "(get one at https://aistudio.google.com/app/apikey)."
-    );
-  }
-
-  cachedAgProjectId = projectId;
-  cachedAgTierId = resolvedTier;
-  cachedAgTierName = loadRes.paidTier?.name || null;
-  log(`[Antigravity] User set up, project: ${projectId}, tier: ${resolvedTier}`);
-  return { projectId, tierId: resolvedTier };
-}
-
-/** Short display name for the Antigravity tier (status line). */
-export function getAntigravityTierDisplayName(): string {
-  if (!cachedAgTierName && !cachedAgTierId) return "Antigravity";
-  const id = cachedAgTierId || "";
-  if (id.includes("ultra")) return "Antigravity Ultra";
-  if (id.includes("pro")) return "Antigravity Pro";
-  if (id === "free-tier") return "Antigravity Free";
-  return cachedAgTierName || "Antigravity";
-}
-
-/** The Antigravity `fetchAvailableModels` response (only the fields we read). */
-interface FetchAvailableModelsResponse {
-  /** The served set — the KEYS of this dict ARE the served model ids. */
-  models?: Record<string, unknown>;
-  /** Backend-provided default model id (e.g. "gemini-3.6-flash-high"). */
-  defaultAgentModelId?: string;
-}
-
-/** The live served-set + default id for the Antigravity account. */
-export interface AntigravityServedModels {
-  servedIds: string[];
-  defaultId: string | null;
-}
-
-// Cached separately from the go@ served set — this comes from a DIFFERENT
-// endpoint (fetchAvailableModels vs retrieveUserQuota) and returns the suffixed
-// served ids that the Antigravity backend actually accepts.
-let agServedCache: AntigravityServedModels | null = null;
-let agServedCacheAt = 0;
-
-/**
- * Discover which model ids this account's Antigravity subscription serves — LIVE.
- *
- * `fetchAvailableModels` is the authoritative per-subscription served-set
- * endpoint: the KEYS of its `models` dict ARE the served ids (already carrying
- * their reasoning-tier suffix, e.g. `gemini-3.6-flash-high`), and
- * `defaultAgentModelId` is the backend's own default. This REPLACES the quota-
- * bucket inference used on the go@ path (buckets don't carry the suffixed ids).
- *
- * Cached for SERVED_MODELS_TTL_MS (the served set changes on the daily cadence,
- * not per request). Degrades gracefully to `{ servedIds: [], defaultId: null }`
- * on any error, so the transport's 404 handling still functions.
- */
-export async function getServedAntigravityModels(
-  accessToken: string,
-  projectId: string,
-  opts?: { force?: boolean }
-): Promise<AntigravityServedModels> {
-  const now = Date.now();
-  if (!opts?.force && agServedCache && now - agServedCacheAt < SERVED_MODELS_TTL_MS) {
-    return agServedCache;
-  }
-  try {
-    const res = await fetch(`${CODE_ASSIST_API_BASE}:fetchAvailableModels`, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-        "Content-Type": "application/json",
-        "User-Agent": buildAntigravityUserAgent(),
-      },
-      // NOTE: the body is `{ project }` — NOT `{ metadata }` (that 400s here).
-      body: JSON.stringify({ project: projectId }),
-    });
-    if (res.ok) {
-      const data = (await res.json()) as FetchAvailableModelsResponse;
-      const servedIds = data.models ? Object.keys(data.models) : [];
-      const defaultId =
-        typeof data.defaultAgentModelId === "string" ? data.defaultAgentModelId : null;
-      if (servedIds.length > 0) {
-        agServedCache = { servedIds, defaultId };
-        agServedCacheAt = now;
-        return agServedCache;
-      }
-    } else {
-      log(`[Antigravity] fetchAvailableModels failed: ${res.status}`);
-    }
-  } catch (err) {
-    log(`[Antigravity] fetchAvailableModels error: ${err}`);
-  }
-  if (agServedCache) return agServedCache;
-  return { servedIds: [], defaultId: null };
-}
-
-/** Test seam: clear the Antigravity served-models cache between tests. */
-export function _resetAntigravityServedModelsCache(): void {
-  agServedCache = null;
-  agServedCacheAt = 0;
-}
