+++ packages/cli/src/telemetry.ts
@@ -48,7 +48,6 @@ const KNOWN_PUBLIC_HOSTS = new Set([
 const PUBLIC_PROVIDERS = new Set([
   "openrouter",
   "gemini",
-  "gemini-codeassist",
   "openai",
   "vertex",
   "ollamacloud",
